+++ metric/instrument/syncfloat64.go
@@ -18,12 +18,18 @@ import (
 	"context"
 
 	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/metric/embedded"
 )
 
 // Float64Counter is an instrument that records increasing float64 values.
 //
-// Warning: methods may be added to this interface in minor releases.
+// Warning: Methods may be added to this interface in minor releases. See
+// [go.opentelemetry.io/otel/metric] package documentation on API
+// implementation for information on how to set default behavior for
+// unimplemented methods.
 type Float64Counter interface {
+	embedded.Float64Counter
+
 	// Add records a change to the counter.
 	Add(ctx context.Context, incr float64, attrs ...attribute.KeyValue)
 }
@@ -64,8 +70,13 @@ type Float64CounterOption interface {
 // Float64UpDownCounter is an instrument that records increasing or decreasing
 // float64 values.
 //
-// Warning: methods may be added to this interface in minor releases.
+// Warning: Methods may be added to this interface in minor releases. See
+// [go.opentelemetry.io/otel/metric] package documentation on API
+// implementation for information on how to set default behavior for
+// unimplemented methods.
 type Float64UpDownCounter interface {
+	embedded.Float64UpDownCounter
+
 	// Add records a change to the counter.
 	Add(ctx context.Context, incr float64, attrs ...attribute.KeyValue)
 }
@@ -107,8 +118,13 @@ type Float64UpDownCounterOption interface {
 // Float64Histogram is an instrument that records a distribution of float64
 // values.
 //
-// Warning: methods may be added to this interface in minor releases.
+// Warning: Methods may be added to this interface in minor releases. See
+// [go.opentelemetry.io/otel/metric] package documentation on API
+// implementation for information on how to set default behavior for
+// unimplemented methods.
 type Float64Histogram interface {
+	embedded.Float64Histogram
+
 	// Record adds an additional value to the distribution.
 	Record(ctx context.Context, incr float64, attrs ...attribute.KeyValue)
 }