+++ sdk/metric/instrument.go
@@ -20,6 +20,7 @@ import (
 	"fmt"
 
 	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/metric/embedded"
 	"go.opentelemetry.io/otel/metric/instrument"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
 	"go.opentelemetry.io/otel/sdk/metric/aggregation"
@@ -171,6 +172,13 @@ type streamID struct {
 
 type instrumentImpl[N int64 | float64] struct {
 	aggregators []internal.Aggregator[N]
+
+	embedded.Float64Counter
+	embedded.Float64UpDownCounter
+	embedded.Float64Histogram
+	embedded.Int64Counter
+	embedded.Int64UpDownCounter
+	embedded.Int64Histogram
 }
 
 var _ instrument.Float64Counter = (*instrumentImpl[float64])(nil)
@@ -213,6 +221,10 @@ type observablID[N int64 | float64] struct {
 type float64Observable struct {
 	instrument.Float64Observable
 	*observable[float64]
+
+	embedded.Float64ObservableCounter
+	embedded.Float64ObservableUpDownCounter
+	embedded.Float64ObservableGauge
 }
 
 var _ instrument.Float64ObservableCounter = float64Observable{}
@@ -228,6 +240,10 @@ func newFloat64Observable(scope instrumentation.Scope, kind InstrumentKind, name
 type int64Observable struct {
 	instrument.Int64Observable
 	*observable[int64]
+
+	embedded.Int64ObservableCounter
+	embedded.Int64ObservableUpDownCounter
+	embedded.Int64ObservableGauge
 }
 
 var _ instrument.Int64ObservableCounter = int64Observable{}